make normal NFS read/write go direct-IO
This avoids mixing direct-IO and normal I/O on an opened file whose flag
is O_DIRECT (unless we close and reopen with desired flag):
1. mixing is not recommended (man 2 open);
2. in FSAL_SECNFS, we have workaround to do write-then-write_plus. So we
do not want to mix at this stage.

Under normal circumstances (in secnfs), the normal buffered I/O is only used
for header read/write (though we may use dix IO for header in the future
to ensure header's integrity).

Note that xfstest may create/list empty files intensely, in which we have to
write/read file header. So this direct-IO may hurt test result.