Allocate heap memory for READ_PLUS4res.rpr_resok4.rpr_contents_val
struct io_info contains a "contents" which is allocated on the stack in
nfs4_op_read_plus(); we could not set "rpr_contents_val" to "&info.contents".
We need to allocate a "contents" from heap in "fill_read_plus_res".  Ownership
of the buffers attached to "info.contents" will be transfered to the newly
allocated "cond".

All allocated heap memory is released in nfs4_op_read_plus_Free().

This patch also contains a trivial bugfix in get_protection_type4().